Update actiosn footer stocktake detail view
TODO fix bug where actions footer disappears when editmodal is opened
