+++ client/packages/inventory/src/Stocktake/DetailView/Toolbar.tsx
@@ -2,9 +2,6 @@ import React from 'react';
 import {
   AppBarContentPortal,
   Grid,
-  DropdownMenu,
-  DropdownMenuItem,
-  DeleteIcon,
   useTranslation,
   BufferedTextInput,
   useBufferState,
@@ -15,22 +12,14 @@ import {
   DateUtils,
   Alert,
   useUrlQuery,
-  RewindIcon,
-  useEditModal,
-  useNotification,
-  ArrowRightIcon,
 } from '@openmsupply-client/common';
 import { useStocktake } from '../api';
-import { ReduceLinesToZeroConfirmationModal } from './ReduceLinesToZeroModal';
-import { ChangeLocationConfirmationModal } from './ChangeLocationModal';
 
 export const Toolbar = () => {
-  const { info } = useNotification();
   const isDisabled = useStocktake.utils.isDisabled();
   const t = useTranslation();
   const { isLocked, stocktakeDate, description, update } =
     useStocktake.document.fields(['isLocked', 'description', 'stocktakeDate']);
-  const onDelete = useStocktake.line.deleteSelected();
   const [descriptionBuffer, setDescriptionBuffer] = useBufferState(description);
 
   const infoMessage = isLocked
@@ -44,33 +33,6 @@ export const Toolbar = () => {
   const setItemFilter = (itemFilter: string) =>
     updateQuery({ itemCodeOrName: itemFilter });
 
-  const reduceModal = useEditModal();
-  const changeLocationModal = useEditModal();
-
-  const selectedRows = useStocktake.utils.selectedRows();
-
-  const checkSelected = () => {
-    if (!selectedRows.length) {
-      const selectRowsSnack = info(t('messages.no-lines-selected'));
-      selectRowsSnack();
-      return;
-    }
-    if (isDisabled) {
-      const isLockedSnack = info(t('error.is-locked'));
-      isLockedSnack();
-      return;
-    }
-    return true;
-  };
-
-  const openReduceToZeroModal = () => {
-    if (checkSelected()) reduceModal.onOpen();
-  };
-
-  const openChangeLocationModal = () => {
-    if (checkSelected()) changeLocationModal.onOpen();
-  };
-
   return (
     <AppBarContentPortal sx={{ display: 'flex', flex: 1, marginBottom: 1 }}>
       <Grid
@@ -80,18 +42,6 @@ export const Toolbar = () => {
         flex={1}
         alignItems="flex-end"
       >
-        {reduceModal.isOpen && (
-          <ReduceLinesToZeroConfirmationModal
-            isOpen={reduceModal.isOpen}
-            onCancel={reduceModal.onClose}
-          />
-        )}
-        {changeLocationModal.isOpen && (
-          <ChangeLocationConfirmationModal
-            isOpen={changeLocationModal.isOpen}
-            onCancel={changeLocationModal.onClose}
-          />
-        )}
         <Grid item display="flex" flex={1} flexDirection="column" gap={1}>
           <InputWithLabelRow
             label={t('heading.description')}
@@ -137,25 +87,6 @@ export const Toolbar = () => {
               setItemFilter(newValue);
             }}
           />
-          <DropdownMenu disabled={isDisabled} label={t('label.actions')}>
-            <>
-              <DropdownMenuItem
-                IconComponent={ArrowRightIcon}
-                onClick={openChangeLocationModal}
-              >
-                {t('button.change-location')}
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                IconComponent={RewindIcon}
-                onClick={openReduceToZeroModal}
-              >
-                {t('button.reduce-lines-to-zero')}
-              </DropdownMenuItem>
-            </>
-            <DropdownMenuItem IconComponent={DeleteIcon} onClick={onDelete}>
-              {t('button.delete-lines')}
-            </DropdownMenuItem>
-          </DropdownMenu>
         </Grid>
       </Grid>
     </AppBarContentPortal>
